Move LRUDict to audit module
