+++ app/api/logging/audit.py
@@ -5,12 +5,10 @@
 # https://docs.python.org/3/library/sys.html#sys.addaudithook
 # https://www.python.org/dev/peps/pep-0578/
 #
-
+import collections
 import logging
 import sys
-from typing import Any, Sequence
-
-import api.util.collections
+from typing import Any, Hashable, Sequence
 
 logger = logging.getLogger(__name__)
 
@@ -95,4 +93,35 @@ def log_audit_event(event_name: str, args: Sequence[Any], arg_names: Sequence[st
     logger.log(AUDIT, event_name, extra=extra)
 
 
-audit_message_count = api.util.collections.LeastRecentlyUsedDict()
+class LeastRecentlyUsedDict(collections.OrderedDict):
+    """A dict with a maximum size, evicting the least recently written key when full.
+
+    Getting a key that is not present returns a default value of 0.
+
+    Setting a key marks it as most recently used and removes the oldest key if full.
+
+    May be useful for tracking the count of items where limited memory usage is needed even if
+    the set of items can be unlimited.
+
+    Based on the example at
+    https://docs.python.org/3/library/collections.html#ordereddict-examples-and-recipes
+    """
+
+    def __init__(self, maxsize: int = 128, *args: Any, **kwargs: Any) -> None:
+        self.maxsize = maxsize
+        super().__init__(*args, **kwargs)
+
+    def __getitem__(self, key: Hashable) -> int:
+        if key in self:
+            return super().__getitem__(key)
+        return 0
+
+    def __setitem__(self, key: Hashable, value: int) -> None:
+        if key in self:
+            self.move_to_end(key)
+        super().__setitem__(key, value)
+        if self.maxsize < len(self):
+            self.popitem(last=False)
+
+
+audit_message_count = LeastRecentlyUsedDict()
