--- app/api/util/collections/__init__.py
@@ -1,5 +1,0 @@
-#
-# Container datatypes and utilities.
-#
-
-from .dict import LeastRecentlyUsedDict  # noqa: F401
